requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
ale-py==0.10.2
alembic==1.15.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
bottle==0.13.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
deepdiff==8.4.2
dm-env==1.6
dm-env-rpc==1.1.6
dm-tree==0.1.9
enum34==1.1.10
Farama-Notifications==0.0.4
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
gymnasium==1.1.1
h5py==3.13.0
idna==3.10
immutabledict==4.2.1
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
loguru==0.7.3
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mock==5.2.0
mpmath==1.3.0
mpyq==0.2.5
multidict==6.4.3
namex==0.0.8
narwhals==1.34.1
networkx==3.4.2
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
optuna==4.2.1
optuna-dashboard==0.18.0
orderly-set==5.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
portpicker==1.6.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
PySC2==4.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
s2clientprotocol==5.0.14.93333.0
s2protocol==5.0.14.93333.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sk-video==1.1.10
SQLAlchemy==2.0.40
stable_baselines3==2.6.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.6.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.20.0
